% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpwas.R
\name{gpwas}
\alias{gpwas}
\title{Genome-Phenome Wide Association Study}
\usage{
gpwas(ingeno, inpheno, inpc, g, gp, gv, R = num, pc = 3,
  selectIn = 0.01, selectOut = 0.01)
}
\arguments{
\item{ingeno}{Input genotype file name/directory.}

\item{inpheno}{Input phenotype file name/directory.}

\item{inpc}{Input folder with PCA parsed population structure covariance. If n number of chromosomes, n number of separate files should be included, as SNPs on each chromosome is excluded for performing PCA once.}

\item{g}{A list of specific gene that needs to analysis. By default the model will run for all of genes detected in the input genotype file.}

\item{gp}{Output file name/directory for selected phenotypes with every gene as well as p value of each selected phenotype.}

\item{gv}{Output file name/directory of terminated p value for each gene.}

\item{R}{Number of iteration for scanning all of input phenotypes with one specific gene.}

\item{pc}{Number of principle components that need to be included as covariances. Default is 3.}

\item{selectIn}{P value threshold for considering a phenotype is significant to be included into the model. Default is 0.01.}

\item{selectOut}{P value threshold for considering a phenotype is significant to be removed from the model. Default is 0.01.}
}
\value{
Two files with significance level per gene and selected phenotypes per gene
}
\description{
Genome-Phenome Wide Association Study
}
\examples{
gpwas(ingeno,inpheno,inpc,g,gp,gv,R=35)
}
